added Password change in UserDetails and LinearLayout structuring

+++ app/src/main/java/com/example/projetofinal_smartcoffee/Database/UserDatabase.java
@@ -16,7 +16,7 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-public class UserDatabase extends Contexter implements IDatabase {
+public class UserDatabase extends Database {
 
     public UserDatabase(String dbName) {
         this.name = dbName;
@@ -26,8 +26,6 @@ public class UserDatabase extends Contexter implements IDatabase {
         setContext(ctx);
         this.name = name;
 
-        // Carregar Users
-        //getAll();
     }
 
     public void open() {
@@ -63,6 +61,10 @@ public class UserDatabase extends Contexter implements IDatabase {
         return pw;
     }
 
+    /**
+     * Adiciona um user a esta db
+     * @param u
+     */
     public void addUser(User u) {
         if(!isOpen) {
             open();
@@ -77,25 +79,101 @@ public class UserDatabase extends Contexter implements IDatabase {
 
     }
 
+    public void updateUser(User u, String newName) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
+        //super.updateAttribute("users", u.getName(), newName);
+        db.execSQL(String.format("UPDATE users SET name='%s' WHERE name='%s'",
+                    newName, u.getName()));
+    }
+
+    public void updateMail(User u, String newMail) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
+        //super.updateAttribute("users", u.getName(), newName);
+        db.execSQL(String.format("UPDATE users SET mail='%s' WHERE mail='%s'",
+                newMail, u.getMail()));
+    }
+
+    public void updatePassword(User u, String password) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
+        db.execSQL(String.format("UPDATE users SET pass='%s' WHERE pass='%s'",
+                password, u.getPass()));
+    }
+
+    /**
+     * Remove um user da db
+     * @param u
+     */
     public void removeUser(User u) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
         db.delete("users", "id = ?", new String[]{String.valueOf(u.getID())});
     }
 
+    /**
+     * Bloqueia o user da db, removendo a possibilidade de login
+     * @param u
+     */
     public void blockUser(User u) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
         if(!isUserBlocked(u.getName())) {
             // Bloquear user
             db.execSQL(String.format("UPDATE users SET status=%d WHERE id=%d", USERSTATE_BLOCKED, u.getID()));
         }
     }
 
+    /**
+     * Desbloqueia o user da db
+     * @param u
+     */
     public void unblockUser(User u) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
         if(isUserBlocked(u.getName())) {
             // Desbloquear user
             db.execSQL(String.format("UPDATE users SET status=%d WHERE id=%d", USERSTATE_NORMAL, u.getID()));
         }
     }
 
+    /**
+     * Permite alterar o tipo de user, este pode ser cliente ou admin
+     * @param u Utilizador
+     * @param type Tipo do utilizador
+     */
     public void setUserType(User u, UserType type) {
+        if(!userExists(u.getName())) {
+            throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
+        }
         int typeValue = USERTYPE_NORMAL;
         switch(type) {
             case Normal: typeValue = USERTYPE_NORMAL; break;
@@ -104,10 +182,13 @@ public class UserDatabase extends Contexter implements IDatabase {
         db.execSQL(String.format("UPDATE users SET type=%s WHERE id=%d", typeValue, u.getID()));
     }
 
-    public User getUserByID(int id) {
-        return null;
-    }
 
+    /**
+     * Retorna um user a partir do seu username de registo
+     * throws UserNotFoundException se o utilizador nao existir
+     * @param name
+     * @return
+     */
     public User getUserByName(String name) {
         for(User u : getAll()) {
             if(u.getName().equals(name)) {
@@ -117,10 +198,20 @@ public class UserDatabase extends Contexter implements IDatabase {
         throw new UserNotFoundException();
     }
 
+    /**
+     * Retorna o estado de conta do user, se esta bloqueado ou nao a partir de um objeto
+     * @param u
+     * @return
+     */
     public boolean isUserBlocked(User u) {
         return isUserBlocked(u.getName());
     }
 
+    /**
+     * Retorna o estado de conta do user, se esta bloqueado ou nao a partir do seu username
+     * @param username
+     * @return
+     */
     public boolean isUserBlocked(String username) {
         int status = USERSTATE_NORMAL;
         Cursor c = db.rawQuery(String.format("SELECT * FROM users WHERE name='%s';", username),null);
@@ -132,6 +223,11 @@ public class UserDatabase extends Contexter implements IDatabase {
         return status == USERSTATE_BLOCKED;
     }
 
+    /**
+     * Verifica se o user existe na db
+     * @param u
+     * @return
+     */
     public boolean userExists(String u) {
         try {
             Cursor c = db.rawQuery(String.format("SELECT name FROM users WHERE name='%s';", u),null);
@@ -161,7 +257,14 @@ public class UserDatabase extends Contexter implements IDatabase {
         return String.format("%s//%s", getContext().getFilesDir().getPath(), name);
     }
 
+    /**
+     * Retorna todos os users na db
+     * @return
+     */
     public List<User> getAll() {
+        if(!isOpen) {
+            open();
+        }
         Cursor c = db.rawQuery("SELECT * FROM users ORDER BY id", null);
         ArrayList<User> users = new ArrayList<>();
 
@@ -178,9 +281,17 @@ public class UserDatabase extends Contexter implements IDatabase {
         return users;
     }
 
+    /**
+     * Verifica se o user se encontra na posicao de admin
+     * @param u
+     * @return
+     */
     public boolean isUserAdmin(User u) {
         if(!userExists(u.getName())) {
             throw new UserNotFoundException();
+        }
+        if(!isOpen) {
+            open();
         }
         Cursor c = db.rawQuery(String.format("SELECT * FROM users WHERE id='%s';", u.getID()),null);
         int type = USERTYPE_NORMAL;
@@ -193,7 +304,7 @@ public class UserDatabase extends Contexter implements IDatabase {
 
     public SQLiteDatabase getHandle() { return db; }
 
-    private SQLiteDatabase db;
+    //private SQLiteDatabase db;
     private String name;
 
     private List<User> users = new ArrayList<>();
